requirements.txt
numpy>=1.26
opencv-python>=4.8
pypdf>=6.0
PyYAML>=6.0
torch>=2.2
torchvision>=0.17
tqdm>=4.66
matplotlib>=3.8
